block go back tx submitted
