+++ apps/wallet-mobile/src/TxHistory/TxHistoryNavigator.tsx
@@ -154,9 +154,17 @@ export const TxHistoryNavigator = () => {
           }}
         />
 
-        <Stack.Screen name="send-submitted-tx" component={SubmittedTxScreen} options={{headerShown: false}} />
+        <Stack.Screen
+          name="send-submitted-tx"
+          component={SubmittedTxScreen}
+          options={{headerShown: false, gestureEnabled: false}}
+        />
 
-        <Stack.Screen name="send-failed-tx" component={FailedTxScreen} options={{headerShown: false}} />
+        <Stack.Screen
+          name="send-failed-tx"
+          component={FailedTxScreen}
+          options={{headerShown: false, gestureEnabled: false}}
+        />
       </Stack.Navigator>
 
       <ModalInfo hideModalInfo={hideModalInfo} visible={modalInfoState}>
