+++ apps/wallet-mobile/src/features/Send/useCases/ConfirmTx/SubmittedTx/SubmittedTxScreen.tsx
@@ -2,12 +2,13 @@ import React from 'react'
 import {StyleSheet, View} from 'react-native'
 
 import {Button, Spacer, Text} from '../../../../../components'
-import {useWalletNavigation} from '../../../../../navigation'
+import {useBlockGoBack, useWalletNavigation} from '../../../../../navigation'
 import {COLORS} from '../../../../../theme'
 import {useStrings} from '../../../common/strings'
 import {SubmittedTxImage} from './SubmittedTxImage'
 
 export const SubmittedTxScreen = () => {
+  useBlockGoBack()
   const strings = useStrings()
   const {resetToTxHistory} = useWalletNavigation()
